fix(node-ui): move PCA client to sibling pca-api.ts (avoid api.ts/api/ dev-server collision)
The extracted PCA client was placed at src/ui/api/pca.ts, but a `src/ui/api/`
directory next to the sibling `src/ui/api.ts` file breaks the vite DEV server: it
serves `/ui/api/pca.ts` with an empty MIME type, the module fails to load, and the
app never mounts. This surfaced only in the node-ui e2e devnet lane
(publishing-conviction-hw-mock.spec.ts → `pca-launch-btn` not visible); the
production rollup build bundles everything, so `build:ui`, tsc, and the unit tests
all passed and hid it — the collision exists only under `dev:ui` per-file serving.

Move the module to a sibling `src/ui/pca-api.ts` (no `api/` directory); the barrel
becomes `export * from './pca-api.js'` and pca-api.ts imports transport from
`./http.js`. Reproduced the failure and verified the fix by driving the vite dev
server with playwright (app mounts, pca-launch-btn visible). tsc zero error-delta,
pca-api 28/28, blast radius 846 pass/0 fail, build:ui exit 0.
